Extract metrics handling into dedicated handler
